Fix integer narrowing conv. (from test), use size_t for arbitrary count.

+++ src/math/hash.cpp
@@ -88,7 +88,7 @@ long_hash hmac_sha512_hash(data_slice data, data_slice key)
 }
 
 bool pkcs5_pbkdf2_hmac_sha512(const std::string& passphrase, 
-    data_slice salt, uint32_t iterations, long_hash& long_hash)
+    data_slice salt, size_t iterations, long_hash& long_hash)
 {
     return pkcs5_pbkdf2(passphrase.c_str(), passphrase.length(),
         salt.data(), salt.size(), long_hash.data(), sizeof(long_hash),
